pluginizes hopline bot
