+++ config-example.json
@@ -1,15 +1,40 @@
 {
   "token": "",
-  "mysql": {
-    "enabled": false,
-    "host": "",
-    "user": "",
-    "password": "",
-    "database": ""
-  },
-  "elasticsearch": {
-    "enabled": false,
-    "host": "localhost:9200",
-    "log": "trace"
+  "plugins": {
+    "roll": {},
+    "convert": {},
+    "bjcp": {},
+    "beers": {},
+    "apero": {},
+    "whisper": {
+      "commands": {
+        "help": "Petite liste non exhaustive\n Pour les douceurs : !roll XdY !café, !bière nomdelabière, !apéro, !arkpéro, !bazoo, !gamb, !alks\nPour les sérieux : !plato2sg degrésplato, !brix2sg degrésbrix, !fermbrix2sg brixoriginal brixactuel\nPour apprendre : !apuglouglou, !bjcp nom du style"
+      }
+    },
+    "text": {
+      "commands": {
+        "bazoo": "Connard",
+        "gamb": "Vive l'Alsace.",
+        "alks": ":candy:.",
+        "arkpéro": "'Il est toujours l'heure de l'arkpéro",
+        "apuglouglou": "Très cher, je te conseille de lire ceci : http://univers-biere.net/pbcourants.php ou à défaut si t'enquérir de solutions te paraît trop long, de taper !bazoo",
+        "café": ":coffee:. Tu veux pas une bière plutôt ? :beer:"
+      }
+    },
+    "welcome": {
+      "message": "Bienvenue sur la Hopline !\nPour que tout se passe bien lis attentivement : http://hopline.fr/bienvenue !\nComme pour tout, il est interdit de parler de choses illégales,\nl'alcool est à consommer avec modération et il est nécessaire d'avoir l'âge légal pour consommer de l'alcool\n(un petit lien dans ta signature incitant les gens à nous rejoindre est un gros plus pour la communauté)"
+    },
+    "mysql": {
+      "enabled": false,
+      "host": "",
+      "user": "",
+      "password": "",
+      "database": ""
+    },
+    "elasticsearch": {
+      "enabled": false,
+      "host": "localhost:9200",
+      "log": "trace"
+    }
   }
 }
