+++ package.json
@@ -22,6 +22,7 @@
     "discord.js": "^11.2.1",
     "elasticsearch": "^14.0.0",
     "html-entities": "^1.2.1",
+    "lodash": "^4.17.5",
     "mysql2": "^1.5.1",
     "string": "^3.3.3",
     "xml2js": "^0.4.19"
